move exporter tests to their own file.
